Rename the files to something more reasonable.
